airflow/requirements.txt
apache-airflow-providers-http
apache-airflow-providers-postgres
requests
anthropic
psycopg2-binary